server/installer/data-downloader/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.23.2
psycopg2-binary>=2.9.9
pydantic==2.9.2
docker>=7.0.0
